% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hazard.R
\name{hazard}
\alias{hazard}
\title{Graph-agnostic hazard functions for BOCPD}
\description{
Hazard functions return the changepoint probability at time \code{t}
conditional on current run length(s) \code{run_length}.
}
