% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hazard.R
\name{hazard_constant}
\alias{hazard_constant}
\title{Constant hazard function}
\usage{
hazard_constant(p)
}
\arguments{
\item{p}{Scalar changepoint probability in \verb{[0,1]}.}
}
\value{
A hazard function \code{h(run_length, t, ...)} returning a numeric
vector of probabilities (same length as \code{run_length}).
}
\description{
Constant hazard function
}
\examples{
h <- hazard_constant(0.01)
h(0:5, t = 10)

}
