% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hazard.R
\name{hazard_geometric}
\alias{hazard_geometric}
\title{Geometric hazard (constant changepoint probability)}
\usage{
hazard_geometric(p)
}
\arguments{
\item{p}{Scalar changepoint probability in \verb{[0,1]}.}
}
\value{
A hazard function.
}
\description{
This is an alias for \code{hazard_constant()}, included for clarity because
the induced segment length distribution is geometric.
}
