% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hazard.R
\name{hazard_piecewise_time}
\alias{hazard_piecewise_time}
\title{Piecewise-constant hazard over time}
\usage{
hazard_piecewise_time(times, p)
}
\arguments{
\item{times}{Integer vector of time cutpoints (strictly increasing).
Hazard \code{p[i]} is used for \code{t <= times[i]} and the last value is used
after the final cutpoint.}

\item{p}{Numeric vector of probabilities in \verb{[0,1]} of length \code{length(times)+1}.}
}
\value{
A hazard function.
}
\description{
The hazard depends on time \code{t} only via user-supplied cutpoints.
}
\examples{
h <- hazard_piecewise_time(times = c(50, 100), p = c(0.01, 0.05, 0.02))
h(0:3, t = 25)
h(0:3, t = 80)

}
