% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_length.R
\name{log_sum_exp}
\alias{log_sum_exp}
\title{Stable log-sum-exp}
\usage{
log_sum_exp(x)
}
\arguments{
\item{x}{Numeric vector (may contain -Inf).}
}
\value{
A single number: log(sum(exp(x))) computed stably.
}
\description{
Stable log-sum-exp
}
\keyword{internal}
