% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_base.R
\name{model_update}
\alias{model_update}
\title{Update model state after observing data}
\usage{
model_update(model, state, x_t)
}
\arguments{
\item{model}{Model object.}

\item{state}{Current model state.}

\item{x_t}{New observation.}
}
\value{
Updated model state.
}
\description{
Update model state after observing data
}
