% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_length.R
\name{normalize_log_probs}
\alias{normalize_log_probs}
\title{Normalize log-probabilities}
\usage{
normalize_log_probs(logp)
}
\arguments{
\item{logp}{Numeric vector of unnormalized log-probabilities.}
}
\value{
Numeric vector of normalized log-probabilities (log-scale),
such that sum(exp(out)) == 1 (up to floating point error).
}
\description{
Normalize log-probabilities
}
\keyword{internal}
